edited heater

+++ shared/src/main/scala/poolmate/Entity.scala
@@ -87,7 +87,7 @@ final case class TimerSetting(id: Long = 0,
 
 final case class Heater(id: Long = 0,
                         poolId: Long = 0,
-                        installed: Int = 0,
+                        installed: Long = 0,
                         model: String = "",
                         cost: Int = 0) extends Entity:
   def display = installed.toString
